fixing some issues, and continue building the app.

+++ app/src/main/java/com/ahmad/houserenovationapp/utils/HelperFunctions.java
@@ -21,7 +21,7 @@ public class HelperFunctions {
         locationB.setLongitude(lon2);
 
         // Calculate the distance in meters
-        double distance = (double)locationA.distanceTo(locationB);
+        double distance = locationA.distanceTo(locationB);
 
         return (int) Math.ceil(distance);
     }
